requirements.txt
mysql~=0.0.3
mysql-connector-python~=8.0.33
fastapi~=0.109.1
pydantic~=2.3.0
passlib~=1.7.4
python-jose~=3.3.0
bcrypt~=4.1.3
python-multipart~=0.0.9
uvicorn~=0.29.0
sqlalchemy~=2.0.3
aiomysql~=0.2.0
pymysql~=1.1.0
cryptography